Force output of BasePortMapper.ports_to_inds to have dtype == int64 to prevent it from returning an empty array with dtype == float64 when passed a selector that doesn't corresp to any ports.

+++ neurokernel/plsel.py
@@ -2020,7 +2020,8 @@ class BasePortMapper(object):
             Integer indices of ports comprised by selector. 
         """
 
-        return self.sel.select(self.portmap, selector).dropna().values
+        return self.sel.select(self.portmap,
+                    selector).dropna().astype(np.int64).values
 
     def get_map(self, selector):
         """
